// partner-search/frontend/src/app/api/model/telecommunication-dto.ts
/**
 * traefik-microstack API Gateway
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {CommunicationTypeDto} from './communication-type-dto';


/**
 * Telecommunication information including phone numbers and email addresses
 */
export interface TelecommunicationDto {
  /**
   * Unique telecommunication identifier - primary key. Optional for new telecommunications (auto-generated), required for updates.
   */
  telecommunicationId?: number;
  /**
   * Telecommunication type identifier
   */
  telecommunicationType?: number;
  /**
   * Address ID - foreign key to AddressEntity
   */
  addressId?: number;
  /**
   * Person ID - foreign key to ContactPersonEntity
   */
  personId?: number;
  /**
   * Country calling code
   */
  countryCode?: string;
  /**
   * Area/city code
   */
  areaCode?: string;
  /**
   * Phone number
   */
  phoneNumber?: string;
  /**
   * Email address
   */
  email?: string;
  /**
   * Last modification date
   */
  modificationDate?: string;
  /**
   * ID of person who made the last modification
   */
  modifierId?: string;
  communicationType?: CommunicationTypeDto;
}